refactor(agent-runtime): close a2a client loop — boot has zero a2a.client imports
- Move RemoteSupport from a2a/client back to a2a (framework bridge, not client impl)
- Merge OpenJiuwenRemoteToolAutoConfiguration into A2aClientAutoConfiguration
  as nested OpenJiuwenRemoteToolConfiguration (one feature = one config entry)
- A2aClientAutoConfiguration manages own daemon thread for card refresh
- RuntimeAutoConfiguration: zero imports from a2a.client package
- AutoConfiguration.imports: 3→2 entries
