+++ agent-runtime/src/main/java/com/huawei/ascend/runtime/engine/a2a/RemoteSupport.java
@@ -1,5 +1,6 @@
-package com.huawei.ascend.runtime.engine.a2a.client;
+package com.huawei.ascend.runtime.engine.a2a;
 
+import com.huawei.ascend.runtime.engine.a2a.client.RemoteAgentInvocationService;
 import java.util.Objects;
 
 /**
