+++ agent-runtime/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,2 +1,2 @@
 com.huawei.ascend.runtime.boot.RuntimeAutoConfiguration
-com.huawei.ascend.runtime.engine.openjiuwen.OpenJiuwenRemoteToolAutoConfiguration
+com.huawei.ascend.runtime.engine.a2a.client.A2aClientAutoConfiguration
