+++ agent-runtime/src/test/java/com/huawei/ascend/runtime/engine/a2a/A2aAgentExecutorTest.java
@@ -17,7 +17,7 @@ import com.huawei.ascend.runtime.engine.spi.AgentExecutionResult;
 import com.huawei.ascend.runtime.engine.spi.AgentRuntimeHandler;
 import com.huawei.ascend.runtime.engine.spi.StreamAdapter;
 import com.huawei.ascend.runtime.engine.a2a.client.RemoteAgentInvocationService;
-import com.huawei.ascend.runtime.engine.a2a.client.RemoteSupport;
+import com.huawei.ascend.runtime.engine.a2a.RemoteSupport;
 import com.huawei.ascend.runtime.engine.spi.TrajectoryDraft;
 import com.huawei.ascend.runtime.engine.spi.TrajectoryEmitter;
 import com.huawei.ascend.runtime.engine.spi.TrajectoryEvent;
